requirements.txt
Flask
python-dotenv
langchain-core
langchain
langchain-community
langchain-experimental
langchain-openai
langchain-text-splitters
openai
faiss-cpu
pypdf
tiktoken
pytube
youtube-transcript-api
pandas
numpy
matplotlib
vercel-wsgi
supabase
storage3
